backend/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dm-tree==0.1.8
exceptiongroup==1.2.1
fastapi==0.110.3
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
gdown==5.1.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
keras==3.2.1
keras-core==0.1.7
kiwisolver==1.4.5
labelImg==1.8.6
libclang==18.1.1
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-macos==2.15.0
termcolor==2.4.0
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.18.1
